refactor: extract webhook tests to separate webhook_tests.rs file
Moves ~57 inline tests from the #[cfg(test)] block in webhook.rs to a
separate src/webhook_tests.rs file, per CLAUDE.md convention. Uses the
#[path = "webhook_tests.rs"] pattern for private access.
